requirements.txt
snowflake-connector-python
pandas
numpy
matplotlib
seaborn
scikit-learn
python-dotenv
